+++ workspaceProvaEmpresa/src/main/java/br/com/contmatic/prova/constants/CidadeConstants.java
@@ -1,0 +1,24 @@
+package br.com.contmatic.prova.constants;
+
+public final class CidadeConstants {
+	
+	public static final int CODIGO_IBGE_TAMNHO_FIXO = 5;
+
+	public static final int NOME_CIDADE_TAMANHO_MINIMO = 3;
+	
+	public static final int NOME_CIDADE_TAMANHO_MAXIMO = 60;
+	
+	public static final int UF_CIDADE_TAMANHO_FIXO = 2;
+
+	// CIDADE
+	
+	public static final String NOME_DEVE_POSSUIR_DE_3_A_60_CARACTERES = "Nome deve possuir de 3 a 60 caracteres!";
+		
+	public static final String CODIGO_IBGE_DEVE_POSSUIR_5_CARACTERES = "Codigo IBGE deve possuir 5 caracteres!";
+
+	public static final String UF_CONTEM_O_NUMERO_DE_DIGITOS_INVALIDO = "UF contem o numero de digitos invalido!";
+		
+	
+	private CidadeConstants() {}
+	
+}
