+++ workspaceProvaEmpresa/src/main/java/br/com/contmatic/prova/constants/ContatoConstants.java
@@ -1,0 +1,27 @@
+package br.com.contmatic.prova.constants;
+
+public final class ContatoConstants {
+	
+		public static final String DDI_BRASIL = "55";
+		
+		public static final int CONDICAO_EMAIL_TAMANHO = 3;
+		
+		public static final int QUANTIDADE_DE_DIGITOS_CELULAR = 13;
+		
+		public static final int PRIMEIRO_CARACTER = 0;
+		
+		public static final String CONTATO_NULO = "Contato invalido!";
+		
+		public static final String CONTATO_EMAIL_INVALIDO = "E-mail invalido!";
+		
+		public static final String CELULAR_NAO_POSSUI_DDI_DO_BRASIL = "Celular não possui DDI do Brasil";
+		
+		public static final String CELULAR_COM_QUANTIDADE_DE_DIGITOS_INVALIDO = "Celular com quantidade de digitos invalido!";
+
+		public static final String CELULAR_INVALIDO_NAO_PODE_CARACTERES_INVALIDOS = "Celular invalido, não pode caracteres invalidos!";
+
+		public static final String EMAIL_NAO_PODE_SER_NULO = "Email não pode ser nulo!";
+		
+		private ContatoConstants() {}
+		
+}
