+++ workspaceProvaEmpresa/src/main/java/br/com/contmatic/prova/constants/CpfConstants.java
@@ -1,0 +1,42 @@
+package br.com.contmatic.prova.constants;
+
+public final class CpfConstants {
+	
+		public static final int ITERACAO2 = 10;
+
+		public static final int ITERACAO1 = 9;
+
+		public static final int CONDICAO_11 = 11;
+
+		public static final int CONDICAO_10 = 10;
+
+		public static final int CPF_SEM_DIGITOS = 9;
+
+		public static final int TAMANHO_CPF = 11;
+
+		public static final int QUANTIDADE_DE_DIGITOS_CPF = 11;
+		
+		public static final int PRIMEIRO_DIGITO_VERIFICADOR = 9;
+		
+		public static final int SEGUNDO_DIGITO_VERIFICADOR = 10;
+		
+		public static final int PESO2 = 11;
+		
+		public static final int PESO_1 = 1;
+		
+		public static final int PESO_10 = 10;
+		
+		public static final int PRIMEIRO_CARACTER = 0;
+		
+		// CPF
+		
+		public static final String CPF_NAO_DEVE_CONTER_LETRAS_E_NEM_CARACTERES_ESPECIAIS = "CPF não deve conter letras e nem caracteres especiais";
+
+		public static final String CPF_INVALIDO = "Cpf invalido!";
+
+		public static final String CPF_INVALIDO_NAO_PODE_SER_NUMERO_SEQUENCIAL = "CPF invalido! CPF não pode ser um numero sequencial!";
+
+		
+		private CpfConstants() {}
+		
+}
