+++ workspaceProvaEmpresa/src/main/java/br/com/contmatic/prova/constants/DepartamentoConstants.java
@@ -1,0 +1,20 @@
+package br.com.contmatic.prova.constants;
+
+public final class DepartamentoConstants {
+	
+		public static final int DEPARTAMENTO_COD_TAMANHO_FIXO = 5;
+
+		public static final int DEPARTAMENTO_NOME_TAMANHO_MAXIMO = 60;
+
+		public static final int DEPARTAMENTO_NOME_TAMANHO_MINIMO = 5;
+		
+		// DEPARTAMENTO
+		
+		public static final String LISTA_DEVE_POSSUIR_DE_2_A_100_FUNCIONARIOS = "Lista deve possuir de 2 a 100 funcionarios!";
+		
+		public static final String COD_DEPARTAMENTO_DEVE_POSSUIR_5_CARACTERES = "Cod Departamento deve possuir 5 caracteres!";
+		
+		
+		private DepartamentoConstants() {}
+		
+}
