+++ workspaceProvaEmpresa/src/main/java/br/com/contmatic/prova/constants/EmpresaConstants.java
@@ -1,0 +1,13 @@
+package br.com.contmatic.prova.constants;
+
+public final class EmpresaConstants {
+
+		public static final int RAZAO_SOCIAL_NOME_TAMANHO_MAXIMO = 60;
+
+		public static final int RAZAO_SOCIAL_NOME_TAMANHO_MINIMO = 5;
+
+		
+		
+		private EmpresaConstants() {}
+		
+}
