+++ workspaceProvaEmpresa/src/main/java/br/com/contmatic/prova/constants/EnderecoConstants.java
@@ -1,0 +1,28 @@
+package br.com.contmatic.prova.constants;
+
+public final class EnderecoConstants {
+
+	public static final int CEP_TAMANHO_FIXO = 8;
+
+	public static final int NUMERO_ENDERECO_TAMANHO_MINIMO = 1;
+
+	public static final int NUMERO_ENDERECO_TAMANHO_MAXIMO = 5;
+
+	public static final int RUA_TAMANHO_MINIMO = 3;
+
+	public static final int RUA_TAMANHO_MAXIMO = 60;
+
+	public static final int TIPO_LOGRADOURO_MIN = 3;
+
+	public static final int TIPO_LOGRADOURO_MAX = 20;
+
+	// ENDERECO
+
+	public static final String CEP_CONTEM_O_NUMERO_DE_DIGITOS_INVALIDO = "CEP contem o numero de digitos invalido!";
+
+	public static final String NUMERO_DA_CASA_CONTEM_A_QUANTIDADE_DE_DIGITOS_INVALIDA = "Numero da casa contem a quantidade de digitos invalida!";
+
+	private EnderecoConstants() {
+	}
+
+}
