+++ workspaceProvaEmpresa/src/main/java/br/com/contmatic/prova/constants/FuncionarioConstants.java
@@ -1,0 +1,25 @@
+package br.com.contmatic.prova.constants;
+
+public final class FuncionarioConstants {
+	
+		public static final int FUNCIONARIO_NOME_TAMANHO_MAXIMO = 60;
+
+		public static final int FUNCIONARIO_NOME_TAMANHO_MINIMO = 5;
+		
+		public static final int ZERO = 0;
+		
+		public static final int IDADE_MINIMA = 18;
+		
+		public static final int TAMANHO_MAX_LISTA_FUNCIONARIO = 100;
+
+		public static final int TAMANHO_MIN_LISTA_FUNCIONARIO = 2;
+		
+		// MENSAGEM
+		
+		public static final String DATA_NULA = "Data invalida!";
+		
+		public static final String PESSOA_COM_IDADE_INSUFICIENTE = "Pessoa não tem idade o suficiente para trabalhar";
+		
+		private FuncionarioConstants() {}
+		
+}
