+++ workspaceProvaEmpresa/src/main/java/br/com/contmatic/prova/constants/UtilsConstants.java
@@ -1,0 +1,44 @@
+package br.com.contmatic.prova.constants;
+
+public final class UtilsConstants {
+	
+	public static final String STRING_INVALIDA = "String invalida!";
+	
+	public static final String NULO = "Não se deve passar um valor nulo";
+	
+	public static final String STRING_INVALIDA_CARACTERES = "String invalida, string contem caracteres invalidos";
+	
+	public static final String STRING_5_A_60 = "String com menos de 5 caracteres ou acima de 60 caracteres";
+	
+	public static final String REGEX_NULA = "Regex nula";
+	
+	public static final int TAMANHO_MIN_LISTA_DEPARTAMENTO = 2;
+	
+	public static final int TAMANHO_MAX_ENDERECO = 20;
+
+	public static final int TAMANHO_MIN_ENDERECO = 1;
+
+	
+	// EMPRESA
+
+	public static final String LISTA_DEVE_POSSUIR_DE_1_A_20_ENDERECOS = "Lista deve possuir de 1 a 20 endereços!";
+
+	public static final String LISTA_DEVE_POSSUIR_DE_2_A_20_DEPARTAMENTOS = "Lista deve possuir de 2 a 20 departamentos!";
+			
+	
+	public static final int TAMANHO_MAX_LISTA_DEPARTAMENTO = 20;
+	
+	public static final int PRIMEIRO_CARACTER = 0;
+	
+	// LISTA VALIDACOES
+	
+	public static final String LISTA_DEVE_CONTER_PELO_MENOS_2_ELEMENTOS_LISTA_INVALIDA = "Lista deve conter pelo menos 2 elementos, lista invalida!";
+	
+	// STRING  VALIDACOES
+	
+	public static final String STRING_INVALIDA_PODE_CONTER_APENAS_NUMEROS = "String invalida, pode conter apenas numeros";
+		
+	public static final String STRING_INVALIDA_PODE_CONTER_APENAS_LETRAS = "String invalida, pode conter apenas letras";
+		
+		
+}
