+++ workspaceProvaEmpresa/src/main/java/br/com/contmatic/prova/endereco/Endereco.java
@@ -1,21 +1,21 @@
-package br.com.contmatic.empresa.endereco;
+package br.com.contmatic.prova.endereco;
 
-import static br.com.contmatic.empresa.constants.Constants.CEP_TAMANHO_FIXO;
-import static br.com.contmatic.empresa.constants.Constants.NUMERO_ENDERECO_TAMANHO_MAXIMO;
-import static br.com.contmatic.empresa.constants.Constants.NUMERO_ENDERECO_TAMANHO_MINIMO;
-import static br.com.contmatic.empresa.constants.Constants.RUA_TAMANHO_MAXIMO;
-import static br.com.contmatic.empresa.constants.Constants.RUA_TAMANHO_MINIMO;
-import static br.com.contmatic.empresa.constants.Constants.TIPO_LOGRADOURO_MAX;
-import static br.com.contmatic.empresa.constants.Constants.TIPO_LOGRADOURO_MIN;
-import static br.com.contmatic.empresa.constants.Mensagens.CEP_CONTEM_O_NUMERO_DE_DIGITOS_INVALIDO;
-import static br.com.contmatic.empresa.constants.Mensagens.NOME_DEVE_POSSUIR_DE_3_A_60_CARACTERES;
-import static br.com.contmatic.empresa.constants.Mensagens.NUMERO_DA_CASA_CONTEM_A_QUANTIDADE_DE_DIGITOS_INVALIDA;
-import static br.com.contmatic.empresa.util.ValidatorUtil.validarCaracteresLetrasEAcentos;
-import static br.com.contmatic.empresa.util.ValidatorUtil.validarCaracteresLetrasENumerosEEspacosEAcentos;
-import static br.com.contmatic.empresa.util.ValidatorUtil.validarCaracteresNumeros;
-import static br.com.contmatic.empresa.util.ValidatorUtil.validarNulo;
-import static br.com.contmatic.empresa.util.ValidatorUtil.validarSeESequencial;
-import static br.com.contmatic.empresa.util.ValidatorUtil.validarTamanhoString;
+import static br.com.contmatic.prova.constants.EnderecoConstants.CEP_TAMANHO_FIXO;
+import static br.com.contmatic.prova.constants.EnderecoConstants.NUMERO_ENDERECO_TAMANHO_MAXIMO;
+import static br.com.contmatic.prova.constants.EnderecoConstants.NUMERO_ENDERECO_TAMANHO_MINIMO;
+import static br.com.contmatic.prova.constants.EnderecoConstants.RUA_TAMANHO_MAXIMO;
+import static br.com.contmatic.prova.constants.EnderecoConstants.RUA_TAMANHO_MINIMO;
+import static br.com.contmatic.prova.constants.EnderecoConstants.TIPO_LOGRADOURO_MAX;
+import static br.com.contmatic.prova.constants.EnderecoConstants.TIPO_LOGRADOURO_MIN;
+import static br.com.contmatic.prova.constants.EnderecoConstants.CEP_CONTEM_O_NUMERO_DE_DIGITOS_INVALIDO;
+import static br.com.contmatic.prova.constants.CidadeConstants.NOME_DEVE_POSSUIR_DE_3_A_60_CARACTERES;
+import static br.com.contmatic.prova.constants.EnderecoConstants.NUMERO_DA_CASA_CONTEM_A_QUANTIDADE_DE_DIGITOS_INVALIDA;
+import static br.com.contmatic.prova.util.ValidatorUtil.validarCaracteresLetrasEAcentos;
+import static br.com.contmatic.prova.util.ValidatorUtil.validarCaracteresLetrasENumerosEEspacosEAcentos;
+import static br.com.contmatic.prova.util.ValidatorUtil.validarCaracteresNumeros;
+import static br.com.contmatic.prova.util.ValidatorUtil.validarNulo;
+import static br.com.contmatic.prova.util.ValidatorUtil.validarSeESequencial;
+import static br.com.contmatic.prova.util.ValidatorUtil.validarTamanhoString;
 
 import java.util.Objects;
