+++ workspaceProvaEmpresa/src/main/java/br/com/contmatic/prova/util/ValidatorUtil.java
@@ -1,17 +1,17 @@
-package br.com.contmatic.empresa.util;
+package br.com.contmatic.prova.util;
 
-import static br.com.contmatic.empresa.constants.Constants.PRIMEIRO_CARACTER;
-import static br.com.contmatic.empresa.constants.Constants.REGEX_ACEITA_LETRAS_COM_ACENTOS;
-import static br.com.contmatic.empresa.constants.Constants.REGEX_ACEITA_LETRAS_E_ESPACOS_E_ACENTOS;
-import static br.com.contmatic.empresa.constants.Constants.REGEX_ACEITA_LETRAS_E_ESPACOS_E_NUMEROS_E_ACENTOS;
-import static br.com.contmatic.empresa.constants.Constants.REGEX_ACEITA_LETRAS_E_NUMEROS_E_ACENTOS;
-import static br.com.contmatic.empresa.constants.Constants.REGEX_ACEITA_NUMEROS;
-import static br.com.contmatic.empresa.constants.Mensagens.NULO;
-import static br.com.contmatic.empresa.constants.Mensagens.REGEX_NULA;
-import static br.com.contmatic.empresa.constants.Mensagens.STRING_5_A_60;
-import static br.com.contmatic.empresa.constants.Mensagens.STRING_INVALIDA_CARACTERES;
-import static br.com.contmatic.empresa.constants.Mensagens.STRING_INVALIDA_PODE_CONTER_APENAS_LETRAS;
-import static br.com.contmatic.empresa.constants.Mensagens.STRING_INVALIDA_PODE_CONTER_APENAS_NUMEROS;
+import static br.com.contmatic.prova.constants.UtilsConstants.PRIMEIRO_CARACTER;
+import static br.com.contmatic.prova.constants.RegexConstants.REGEX_ACEITA_LETRAS_COM_ACENTOS;
+import static br.com.contmatic.prova.constants.RegexConstants.REGEX_ACEITA_LETRAS_E_ESPACOS_E_ACENTOS;
+import static br.com.contmatic.prova.constants.RegexConstants.REGEX_ACEITA_LETRAS_E_ESPACOS_E_NUMEROS_E_ACENTOS;
+import static br.com.contmatic.prova.constants.RegexConstants.REGEX_ACEITA_LETRAS_E_NUMEROS_E_ACENTOS;
+import static br.com.contmatic.prova.constants.RegexConstants.REGEX_ACEITA_NUMEROS;
+import static br.com.contmatic.prova.constants.UtilsConstants.NULO;
+import static br.com.contmatic.prova.constants.UtilsConstants.REGEX_NULA;
+import static br.com.contmatic.prova.constants.UtilsConstants.STRING_5_A_60;
+import static br.com.contmatic.prova.constants.UtilsConstants.STRING_INVALIDA_CARACTERES;
+import static br.com.contmatic.prova.constants.UtilsConstants.STRING_INVALIDA_PODE_CONTER_APENAS_LETRAS;
+import static br.com.contmatic.prova.constants.UtilsConstants.STRING_INVALIDA_PODE_CONTER_APENAS_NUMEROS;
 import static java.lang.String.valueOf;
 import static java.util.regex.Pattern.CASE_INSENSITIVE;
 import static java.util.regex.Pattern.compile;
