+++ workspaceProvaEmpresa/src/test/java/br/com/contmatic/empresa/empresa/CidadeTest.java
@@ -8,7 +8,7 @@ import static org.junit.Assert.assertTrue;
 import org.junit.Before;
 import org.junit.Test;
 
-import br.com.contmatic.empresa.endereco.Cidade;
+import br.com.contmatic.prova.endereco.Cidade;
 
 public class CidadeTest {
