+++ workspaceProvaEmpresa/src/test/java/br/com/contmatic/empresa/empresa/FuncionarioTest.java
@@ -13,9 +13,10 @@ import org.joda.time.DateTime;
 import org.junit.Before;
 import org.junit.Test;
 
-import br.com.contmatic.empresa.contato.Contato;
-import br.com.contmatic.empresa.endereco.Endereco;
-import br.com.contmatic.empresa.funcionario.Funcionario;
+import br.com.contmatic.prova.contato.Contato;
+import br.com.contmatic.prova.empresa.Departamento;
+import br.com.contmatic.prova.empresa.Funcionario;
+import br.com.contmatic.prova.endereco.Endereco;
 
 
 public class FuncionarioTest {
